+++ CHANGELOG.md
@@ -9,3 +9,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add in version field when exporting an instance [#419]
 - Clean up old instance format [#422]
 - Fix some issues with remembering window size and position option [#423]
+- Fix instance installation not cancelling on download exceptions